+++ src/xrCore/stream_reader.cpp
@@ -94,7 +94,7 @@ void CStreamReader::r(void* _buffer, u32 buffer_size)
     }
 
     u8* buffer = (u8*)_buffer;
-    u32 elapsed_in_window = m_current_window_size - (m_current_pointer - m_start_pointer);
+    u32 elapsed_in_window = m_current_window_size - static_cast<u32>(m_current_pointer - m_start_pointer);
 
     do
     {
